recognition-services/requirements.txt
# Dependências do serviço de reconhecimento facial
# Compatível com Python 3.8 - 3.12 (Python 3.13 ainda não é suportado)

# OpenCV para captura e processamento de vídeo
opencv-python>=4.8.0

# Face Recognition (baseado em dlib)
face_recognition>=1.3.0

# Pillow para processamento de imagens (versão compatível com Python 3.13)
Pillow>=10.0.0

# Requests para comunicação com API
requests>=2.31.0

# Python-dotenv para variáveis de ambiente
python-dotenv>=1.0.0

# NumPy (compatível com Python 3.13)
numpy>=1.26.0
